+++ .env
@@ -1,4 +1,4 @@
-export FR_ACCESS_KEY=<YOUR_FR_ACCESS_KEY>
-export FR_SECRET_KEY=<YOUR_FR_SECRET_KEY>
-export MOBILE_NUMBER=<YOUR_MOBILE_NUMBER>
+export FR_ACCESS_KEY=29039577
+export FR_SECRET_KEY=AfOXv5mMAK9mtT4nQKWBBS7dUjiEGu78 
+export MOBILE_NUMBER=12067392634
